Init commit
Banking app includes 3 classes.  The banking class, menu class, and profile class.  A new instance of the banking app is created in the main function and the start function is called through that instance.  Profile instances and menu instances are created where the user has to enter their information, and would have access to their account balance, and would be able to withdraw, deposit, and check balance.